+++ src/renderer/components/Expenses.vue
@@ -126,36 +126,14 @@
         <v-card>
           <v-form ref="expenseForm">
             <v-layout row>
-              <v-flex xs3>
-                <v-menu
-                  tabindex="-1"
-                  ref="dateMenu"
-                  lazy
-                  :close-on-content-click="false"
-                  v-model="showDateMenu"
-                  transition="scale-transition"
-                  offset-y
-                  full-width
-                  :nudge-right="40"
-                  min-width="290px"
-                  :return-value.sync="entryDateStr">
-                  <v-text-field
-                    slot="activator"
-                    label="Date"
-                    v-model="entryDateStr"
-                    prepend-icon="mdi-calendar-range"
-                    readonly
-                    required
-                  ></v-text-field>
-                  <v-date-picker v-model="entryDateStr"
-                    next-icon="mdi-chevron-right"
-                    prev-icon="mdi-chevron-left"
-                    color="green accent-3">
-                    <v-spacer></v-spacer>
-                    <v-btn flat color="error" @click="showDateMenu = false">Cancel</v-btn>
-                    <v-btn color="success" @click="$refs.dateMenu.save(entryDateStr)">OK</v-btn>
-                  </v-date-picker>
-                </v-menu>
+              <v-flex xs2>
+                <v-text-field
+                  label="Date"
+                  v-model="entryDateStr"
+                  prepend-icon="mdi-calendar-range"
+                  required
+                  :rules="rules.date">
+                </v-text-field>
               </v-flex>
               <v-flex xs3>
                 <v-select
@@ -183,7 +161,7 @@
                   v-model="entry.amount">
                 </v-text-field>
               </v-flex>
-              <v-flex xs4>
+              <v-flex xs5>
                 <v-text-field
                   name="notes"
                   label="Notes"
@@ -213,6 +191,7 @@ import ExpenseDB from '../lib/ExpenseDB'
 import Format from '../lib/Format'
 import Constants from '../lib/Constants'
 import Mousetrap from 'Mousetrap'
+import Moment from 'moment'
 
 export default {
   name: 'Expenses',
@@ -409,6 +388,7 @@ export default {
     },
 
     newEntry: function () {
+      this.entryDateStr = Format.formatDate(new Date(), Constants.FORMATS.entryDate)
       this.$refs.categorySelect.$el.focus()
       this.showAddEditSheet = true
     },
@@ -773,7 +753,8 @@ export default {
       showOverbudget: false,
       rules: {
         date: [
-          date => !!date || 'Date is required'
+          date => !!date || 'Date is required',
+          date => (Moment(date, Constants.FORMATS.entryDate, true).isValid()) || 'Format as ' + Constants.FORMATS.entryDate
         ],
         category: [
           category => !!category || 'Category is required'