+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 * Expenses: Include "unbudgeted" categories in Category drop-down.
 * Expenses: Shouldn't need to use +/- sign to indicate Income or Expense
 * Expenses: View only over-budget items
+* Expenses: Make it easier to change the date
 
 ## v1.0.3 (May 31, 2018)
 * Fixed *another* date related issue. The `monthNumberToName` function was returning